// src/app/model/users/users-list.ts
export interface UsersListItem {
    android_device_token: string;
    city_uuid: string;
    company_uuid: string;
    created_dt: string;
    deleted: number
    email_verified: string;
    gender: string;
    id: number
    ios_device_token: string;
    last_login: string;
    location: string;
    locked: string;
    mobile_no: string;
    password_hash: string;
    postal_code: string;
    role: string;
    state_uuid: string;
    status: string;
    updated_dt: string;
    user_email: string;
    user_firstname: string;
    user_image: string;
    user_lastname: string;
    user_lastseen: string;
    user_thumbnail: string;
    uuid: string;
    verified: boolean;
    checked?: boolean;
}